services/enhanced-ingestion-service/requirements.txt
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Google Cloud services
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
google-cloud-vision==3.4.5
google-cloud-speech==2.21.0
google-cloud-firestore==2.13.1
google-cloud-videointelligence==2.11.4

# AI and ML
google-genai==0.3.0

# Document processing
PyPDF2==3.0.1
pymupdf==1.23.8
python-docx==1.1.0
python-pptx==0.6.23
pillow==10.1.0
pytesseract==0.3.10

# Video and audio processing
moviepy==1.0.3
speech-recognition==3.10.0
pydub==0.25.1

# Web scraping and external data
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
trafilatura==1.6.4

# Async task processing
celery==5.3.4
redis==5.0.1

# Utilities
python-multipart==0.0.6
httpx==0.25.2